JACKTECH COMMUNICATION SERVICES
This Is a private Company that deals with all ideas of Technology and communication services with its branches located In Different parts of the country. Its main goal is to provide quality services and skills in the ICT fields, Introduce Technology to the person of any age range between 10 years- 60 years
at the comfort of their working places,homes or any other places. We can operate both online and physically

Here Are Some.OF THE..services OFFERED
1.All Networks Lines activation: .
2. Saf lines registration and activation.
3. TelKom lines registration and activation
4. Airtel registration
5.ACCESSORIES i.e.
6.Cables
7.Earphones
8. Android chargers & cables
9. Internet routers
9. Android covers